Task 1

+++ src/main/java/tk/project/goodsstorage/product/dto/CreateProductDto.java
@@ -4,6 +4,8 @@ import lombok.AccessLevel;
 import lombok.Data;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Data
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class CreateProductDto {
@@ -16,7 +18,7 @@ public class CreateProductDto {
 
     String category;
 
-    Integer price;
+    BigDecimal price;
 
     Long count;
 }
