+++ src/main/java/tk/project/goodsstorage/product/dto/ProductDto.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.util.UUID;
@@ -24,7 +25,7 @@ public class ProductDto {
 
     String category;
 
-    Integer price;
+    BigDecimal price;
 
     Long count;
